Split `BinaryTreeLookup` to multiple files
